feat(all): extracted the community creation dialog to a separate reusable component
